Better sync support
