+++ generator/sbpg/targets/resources/sbp2json-cargo.toml
@@ -11,20 +11,9 @@ version = "(((release)))"
 authors = ["Swift Navigation <dev@swiftnav.com>"]
 edition = "2018"
 
-[dependencies.sbp]
-path = "../sbp"
-features = ["sbp_serde", "json"]
-
 [dependencies]
 structopt = "0.3"
 env_logger = "0.8.2"
-tokio-util = { version = "0.5.1", default-features = false, features = ["compat"] }
-futures = { version = "0.3.8" }
-
-[dependencies.tokio]
-version = "0.3"
-default-features = false
-features = ["rt-multi-thread", "io-util", "io-std"]
 
 [target.'cfg(all(not(windows), not(target_env = "musl")))'.dependencies]
 jemallocator = "0.3"
