+++ rust/sbp2json/Cargo.toml
@@ -18,13 +18,6 @@ features = ["sbp_serde", "json"]
 [dependencies]
 structopt = "0.3"
 env_logger = "0.8.2"
-tokio-util = { version = "0.5.1", default-features = false, features = ["compat"] }
-futures = { version = "0.3.8" }
-
-[dependencies.tokio]
-version = "0.3"
-default-features = false
-features = ["rt-multi-thread", "io-util", "io-std"]
 
 [target.'cfg(all(not(windows), not(target_env = "musl")))'.dependencies]
 jemallocator = "0.3"
