+++ rust/sbp2json/src/bin/json2json.rs
@@ -1,5 +1,7 @@
+use std::io;
+
+use sbp::codec::{converters::blocking::json2json, CompactFormatter, HaskellishFloatFormatter};
 use structopt::StructOpt;
-use tokio::runtime;
 
 #[cfg(all(not(windows), not(target_env = "musl")))]
 #[global_allocator]
@@ -22,7 +24,7 @@ struct Options {
     float_compat: bool,
 }
 
-fn main() {
+fn main() -> sbp::Result<()> {
     let options = Options::from_args();
 
     if options.debug {
@@ -31,17 +33,12 @@ fn main() {
 
     env_logger::init();
 
-    let rt = runtime::Builder::new_multi_thread().build().unwrap();
+    let stdin = io::stdin();
+    let stdout = io::stdout();
 
-    rt.block_on(async {
-        let stdin = sbp2json::stdin();
-        let stdout = sbp2json::stdout();
-
-        if options.float_compat {
-            sbp::codec::json2json(stdin, stdout, sbp::codec::HaskellishFloatFormatter {}).await
-        } else {
-            sbp::codec::json2json(stdin, stdout, sbp::codec::CompactFormatter {}).await
-        }
-    })
-    .unwrap()
+    if options.float_compat {
+        json2json(stdin, stdout, HaskellishFloatFormatter {})
+    } else {
+        json2json(stdin, stdout, CompactFormatter {})
+    }
 }
