--- rust/sbp2json/src/lib.rs
@@ -1,9 +1,0 @@
-pub fn stdin() -> impl futures::AsyncRead {
-    use tokio_util::compat::Tokio02AsyncReadCompatExt;
-    tokio::io::stdin().compat()
-}
-
-pub fn stdout() -> impl futures::AsyncWrite {
-    use tokio_util::compat::Tokio02AsyncWriteCompatExt;
-    tokio::io::stdout().compat_write()
-}
